pyproject.toml
[tool.poetry]
name = "chorse"
version = "0.1.0"
description = ""
authors = ["Luca <luca@bluewhale.kr>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
paramiko = "2.6.0"
Pillow = "^6.2.1"
ffmpeg-python = "^0.2.0"
pdfminer=""

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
